Global functions made methods
A number of global functions relevant to Enemy and Player have been
made methods.

A number of dimensions (size of board, tile, enemy, etc.) have been moved to
an object.

Three initial instances of Enemy are not assigned to global variables anymore,
as it was discovered they persist in memory after they are needed.

Revisions to code comments.