Add lives markers, change fonts
Lives remaining to player are indicated with rectangles on the bottom
left of the screen.  In the future, they may be replaced with smaller images
of player.

Webfont (Roboto) is used for score and "GAME OVER".  Right now, it is made
available through a hack: it is first applied to a <h6> element which is
drawn off-screen.