Add tile marker for win, reformat
When player reaches water, the water tile is replaced by a contrasting image
(saturation and lightness modified in GIMP).  The tile reverts when the player
moves again.  It would be nice to do this (and move the player back) on a time
delay related to 'dt'.

Code has been reformatted using JSTools plugin on Notepad++.