Add timer for marked tiles
When player reaches the water, the tile arrived on will change appearance.
Previously, this change reverted when the player moved again.  Now this
happens after 30 frames.

An attempt has been made to avoid setTimeout, which is not synchronized
with screen redraw.  Since requestAnimationFrame is already in use, it makes
sense to hook into the number of elapsed frames.

This code succeeds, but at the expense of global variables and an opaque
interface.  Probably a Timer class would be easier to use and allow multiple
frame-timed events.

This commit also removes some console display code.